Refactor the Bezier curve drawing
* Move Bezier code to a standalone file
* Statically assert number base I for integer and R for real number (floating-point)
* Rename R3DWorld to RWorld
* Rename unit testing aliases Vector3d and Bezier3d to Vector and Bezier
