+++ cpp/modmesh/universe/universe.hpp
@@ -29,6 +29,7 @@
  */
 
 #include <modmesh/universe/bernstein.hpp>
+#include <modmesh/universe/bezier.hpp>
 #include <modmesh/universe/World.hpp>
 
 namespace modmesh
